% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_transform.R
\name{z_transform}
\alias{z_transform}
\title{Z-transform a variable}
\usage{
z_transform(x)
}
\arguments{
\item{x}{a numeric vector to be z-transformed}
}
\value{
Returns a vector
}
\description{
Z-transform a variable. Mean=0, SD=1. Maintains original distribution.
}
\examples{
z_transform(x)

}
\author{
Luke Pilling
}
